--- src/pages/events/[id]/volunteers.tsx
@@ -1,100 +1,0 @@
-import { useEffect, useState } from "react";
-import { useRouter } from "next/router";
-import { Staff } from "@prisma/client";
-
-import AssigneeBadge from "@/components/Cards/assignee-badge";
-import VolunteerManageForm from "@/components/Forms/volunteer-manage-form";
-import Header, { HeaderProps } from "@/components/Header";
-import Modal from "@/components/Modal";
-import Sidebar from "@/components/sidebar/index";
-import LoadingSpinner from "@/components/UI/loading-spinner";
-import { useEvent } from "@/hooks/events";
-
-export default function Volunteers() {
-  const [volunteers, setVolunteers] = useState<Staff[]>([]);
-  const [headerValues, setHeaderValues] = useState<HeaderProps>();
-  const [showVolunteerModal, setShowVolunteerModal] = useState(false);
-  const {
-    query: { id: eventId }
-  } = useRouter();
-
-  const { data: event } = useEvent(eventId as string);
-  useEffect(() => {
-    if (!event) return;
-    setHeaderValues({
-      name: event.name,
-      location: event.location,
-      startDate: new Date(event.startDate),
-      endDate: new Date(event.endDate),
-      createdBy: event.createdBy
-    });
-
-    const urlParams = new URLSearchParams({ id: event.id });
-
-    fetch(`/api/event/${event.id}/repairers?${urlParams.toString()}`)
-      .then((res) => res.json())
-      .then((data) => setVolunteers(data));
-  }, [event]);
-
-  function manageVolunteer() {
-    setShowVolunteerModal(true);
-  }
-  return (
-    <Sidebar>
-      <main className="ml-80 min-h-screen w-full p-4">
-        {headerValues ? (
-          <>
-            <Header {...headerValues} />
-            <div className="container">
-              <div className="flex flex-row w-auto p-4 text-2xl font-bold text-zinc-400 content-center justify-between">
-                <span>Volunteers ({volunteers.length})</span>
-                <button
-                  className="flex rounded-lg border bg-primary-500 p-2 w-1/5 shadow-md transition duration-150 ease-in-out hover:cursor-pointer hover:bg-primary-300 text-sm active:bg-primary-500 text-white justify-center"
-                  onClick={manageVolunteer}
-                  onKeyDown={manageVolunteer}
-                >
-                  Manage Volunteers
-                </button>
-              </div>
-              <div className="container mx-auto">
-                <div className="flex justify-end"></div>
-              </div>
-              <div className="grid gap-4 p-4 lg:grid-cols-5 ">
-                {volunteers.map((item) => (
-                  <div key={item.id}>
-                    <AssigneeBadge
-                      firstName={item.clerkId ?? item.organisationId}
-                      lastName={item.role ?? ""}
-                    />
-                  </div>
-                ))}
-              </div>
-            </div>
-            {event && (
-              <Modal
-                showModal={showVolunteerModal}
-                setShowPopup={setShowVolunteerModal}
-                height="h-full"
-              >
-                <div className="text-center">
-                  <h1 className="text-xl font-bold">Add / Remove Volunteers</h1>
-                  <div>
-                    <VolunteerManageForm
-                      volunteersArray={volunteers.map(
-                        (volunteer) => volunteer.id
-                      )}
-                      setShowModal={setShowVolunteerModal}
-                      eventProps={event}
-                    />
-                  </div>
-                </div>
-              </Modal>
-            )}
-          </>
-        ) : (
-          <LoadingSpinner className="w-full h-full flex items-center justify-center " />
-        )}
-      </main>
-    </Sidebar>
-  );
-}
